Rename to SemanticService
