--- src/Core/src/Accessibility/AccessibilityService.Standard.cs
@@ -1,6 +1,0 @@
-﻿namespace Microsoft.Maui.Accessibility
-{
-	public class AccessibilityService
-	{
-	}
-}
