--- src/Core/src/Accessibility/AccessibilityService.Windows.cs
@@ -1,6 +1,0 @@
-﻿namespace Microsoft.Maui.Accessibility
-{
-	public class AccessibilityService
-	{
-	}
-}
