+++ src/Core/src/Semantics/SemanticService.Windows.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Maui.Semantics
+{
+	public class SemanticService
+	{
+	}
+}
