+++ src/Core/src/Semantics/SemanticService.iOS.cs
@@ -1,9 +1,9 @@
 ﻿using Foundation;
 using UIKit;
 
-namespace Microsoft.Maui.Accessibility
+namespace Microsoft.Maui.Semantics
 {
-	public class AccessibilityService
+	public class SemanticService
 	{
 		public void SetAnnouncement(string text)
 		{
